[MIG] account_bank_statement_import_save_file to v13
