+++ account_bank_statement_import_save_file/__init__.py
@@ -1,1 +1,2 @@
 from . import models
+from . import wizard
