+++ account_bank_statement_import_save_file/models/__init__.py
@@ -1,2 +1,1 @@
 from . import account_bank_statement
-from . import account_bank_statement_import
